We should be able to handle a class which defines its own native "new" or "initialize" method with some required parameters and some optional parameters (e.g. StringScanner).
Native methods with negative arity, only ever report -1, so we can never know how many required parameters there are. A hacky fix is to keep trying more and more parameters until it works or we give up.

We also had to add an explicit to_str method to MethodInvocationSink to avoid StringScanner raising a TypeError. It's not obvious that this approach is sustainable.